refactored and finished log in, out, refreshing and created decroatores to secure presen t and all future routes
