+++ BackEndFlask/controller/__init__.py
@@ -16,3 +16,4 @@ from controller.Routes import Logout_route
 from controller.Routes import Upload_csv_routes
 from controller.Routes import Refresh_route
 from controller.security import utility
+from controller.security import customDecorators
